+++ test/test_helper.rb
@@ -1,0 +1,2 @@
+# Načíta hlavný Redmine test_helper (z koreňa Redmine).
+require File.expand_path(File.dirname(__FILE__) + '/../../../test/test_helper')